Trainingのインテグレーションテストを追加
