+++ src/main/java/jp/classmethod/sparrow/model/Converter.java
@@ -13,34 +13,16 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package model;
+package jp.classmethod.sparrow.model;
 
 import java.io.IOException;
-
-import lombok.extern.slf4j.Slf4j;
-
-import org.springframework.stereotype.Service;
 
 /**
  * Created by kunita.fumiko on 2017/03/22.
  */
 
-@Slf4j
-@Service
-
-public class Processor {
-	public String run(Converter converter, String request) throws IOException {
-		
-		// これから行う処理を説明する(printDescriptionメソッドを呼び出す)
-		printDescription(converter.getDescription());
-		
-		//文字列の変換処理
-		String dest = converter.convert(request);
-		//結果の出力処理
-		return dest;
-	}
+public interface Converter {
+	String convert(String result) throws IOException;
 	
-	private void printDescription(String description) {
-		log.info(description);
-	}
+	String getDescription();
 }
