+++ src/main/java/jp/classmethod/sparrow/model/Processor.java
@@ -13,7 +13,11 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package model;
+package jp.classmethod.sparrow.model;
+
+import java.io.IOException;
+
+import lombok.extern.slf4j.Slf4j;
 
 import org.springframework.stereotype.Service;
 
@@ -21,17 +25,22 @@ import org.springframework.stereotype.Service;
  * Created by kunita.fumiko on 2017/03/22.
  */
 
+@Slf4j
 @Service
-public class DoublingConverter extends AbstractConverter {
-	@Override
-	protected String computeStringToAppend(int c) {
-		StringBuilder sb = new StringBuilder();
-		sb.append((char) c).append((char) c);
-		return sb.toString();
+
+public class Processor {
+	public String run(Converter converter, String request) throws IOException {
+		
+		// これから行う処理を説明する(printDescriptionメソッドを呼び出す)
+		printDescription(converter.getDescription());
+		
+		//文字列の変換処理
+		String dest = converter.convert(request);
+		//結果の出力処理
+		return dest;
 	}
 	
-	//処理説明
-	public String getDescription() {
-		return "文字をダブらせます";
+	private void printDescription(String description) {
+		log.info(description);
 	}
 }
